requirements.txt
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy
Flask-WTF==1.2.2
python-dotenv==1.1.0
# SQLAlchemy==2.0.32
Werkzeug==3.1.3
Jinja2==3.1.6
msal==1.32.0
requests>=2.31.0
google-auth-oauthlib==1.2.2
google-api-python-client==2.171.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
azure-core>=1.30.0
azure-ai-inference==1.0.0b9
flake8==7.2.0
pyodbc==5.1.0
alembic==1.13.0

# Development dependencies
black==25.1.0
mypy==1.16.0
isort==5.13.2